Task 13. JPA/Hibernate database integration with transactional DAO layer

+++ HotelManagementSystem/hotel-service/src/main/java/ru/senla/hotel/service/exception/service/InvalidServicePriceException.java
@@ -1,7 +1,9 @@
 package ru.senla.hotel.service.exception.service;
 
+import java.math.BigDecimal;
+
 public class InvalidServicePriceException extends ServiceException {
-    public InvalidServicePriceException(double price) {
+    public InvalidServicePriceException(BigDecimal price) {
         super("Invalid service price: " + price + ". Price must be non-negative.");
     }
 }
